+++ pulsar/lib/__init__.py
@@ -1,15 +1,5 @@
-#First try local
-try:
-    from ._pulsar import *
-    hasextensions = True
-except ImportError: #pragma    nocover
-    # Try Global
-    try:
-        from _pulsar import *
-        hasextensions = True
-    except ImportError:
-        hasextensions = False
-        from pulsar.utils.httpurl import HttpParser
+hasextensions = False
+from pulsar.utils.httpurl import HttpParser
 
 Http_Parser = HttpParser